requirements.txt
asgiref==3.4.1
astroid==2.8.2
autopep8==1.5.7
bandit==1.7.0
colorama==0.4.4
coverage==6.0.1
Django==3.2.8
django-cors-headers==3.10.0
django-countries==7.2.1
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.24
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pbr==5.6.0
Pillow==8.3.2
platformdirs==2.4.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.11.1
python-decouple==3.5
pytz==2021.3
PyYAML==5.4.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.2
stevedore==3.4.0
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
